pyproject.toml
[project]
name = "crosscoders"
version = "0.1.0"
description = "General training of cross-layer transcoders on LLMs"
authors = [
    {name = "neale",email = "nealeratzlaff@gmail.com"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "transformers",
    "einops",
    "pyyaml",
    "torch",
    "numpy",
    "tqdm",
    "datasets"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
